refactor: change interfaces folder by types since it's more generic
