+++ src/app/components/landing/list/card/card.component.ts
@@ -1,8 +1,8 @@
 
 import { Component, inject, input, OnInit, signal } from '@angular/core';
 import { CapitalizePipe } from '../../../../pipes/capitalize.pipe';
-import { ComponentInfo } from '../../../../interfaces/component-info';
-import { DynamicCardInfo, PartialDynamicCardInfo } from '../../../../interfaces/dynamic-card';
+import { ComponentInfo } from '../../../../types/component-info';
+import { DynamicCardInfo, PartialDynamicCardInfo } from '../../../../types/dynamic-card';
 import { CardService } from '../../../../services/card.service';
 import { RouterLink } from '@angular/router';
 import { TimePipe } from '../../../../pipes/time.pipe';
