+++ src/app/types/component-info.ts
