+++ src/app/types/component-min-info.ts
