+++ src/app/types/dynamic-card.ts
